Update roi functions with correct date format
Both functions were updated and now they use the
static mehod _format_date to format the %d/%m/%Y
string format date to the timestamp format.

Additionaly, now get_periodic_roi has a from_date argument
and a to_date optional argument.

For the cumulative_period_roi, to_date was made optional and
it was added a period argument in the same sense as in the
periodic roi function.