Fix commulative roi function
Fixed a problem with the resample of the data.

Previously we first computed the accumulated ROI and then resampled.

Now we are doing the inverse. With this we guarantee that
the first period will always be zero. We won't get the variation from,
the initial period to the  resample date as our initial value.