Fix in the periodic roi index interval
Now we don't exclude the first value of the series.
We can't compute the roi for the first period it is a NaN value
To keep conformity with the accumulated ROI we will keep the first val.
With this but series will have the same index if they are genereted
with the same to_date, from_date and period. This is really important
for the plot function that plot both series in two subplots with shared
x axis